Fixed the typos, added benchmark details, and added missing license headers. Please let me know if anything else needs attention
